requirements.txt
absl-py==2.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.4
fonttools==4.51.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.2
gym==0.26.2
gym-notices==0.0.8
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==5.26.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pygame==2.5.2
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
qdldl==0.1.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
scs==3.2.3
seaborn==0.12.2
simpy==4.0.1
six==1.16.0
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
threadpoolctl==3.1.0
torch==2.0.0
torch-tb-profiler==0.4.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
